requirements.txt
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
datasets==4.1.1
debugpy==1.8.17
decorator==5.2.1
deepspeed==0.17.6
diffusers @ git+https://github.com/huggingface/diffusers.git@4c723d8ec318b5fd266afaf14ba37afdefc967df
dill==0.4.0
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
ftfy==6.3.1
grpcio==1.75.1
h11==0.16.0
hf-xet==1.1.10
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
ipykernel==7.0.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jupyter_client==8.6.3
jupyter_core==5.9.1
lightning-utilities==0.15.2
Markdown==3.9
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.17.1
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.5.0
prodigyopt==1.1.2
prompt_toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==21.0.0
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.5
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.9.18
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
sympy==1.14.0
taming-transformers-rom1504==0.0.6
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tokenizers==0.22.1
torch==2.8.0
torchao==0.10.0
torchaudio==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
triton==3.4.0
trl==0.19.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.20.1
zipp==3.23.0
